Initial commit, working version
This adds a working version of the pymato timer, which is pre-set
to have 4 iterations of 2 phases (work and rest). The Timer class
does not include hard-coded values, but the calling method does.

There are currently no keybindings used, and the timer cannot be
controlled or exited nicely once started.