Remove default phase list and improve displaying messages.
The default phase list has been removed, and phases can no longer be
passed into the constructor. Instead phases must be explicitly added
once the Timer object has been created. This is with a view to providing
a single method for adding phases which can be used easily when reading
the configuration from a file.

The displaying of messages has been refactored to make some use of the
curses windows. Class methods have also been created to simplify the
calls needed to update the strings, as well as to allow their position
to be more easily updated in the future.